requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
pydantic-settings
python-dotenv